+++ include/superformula_scene.h
@@ -19,7 +19,7 @@ namespace so
 			float m_lastEval[POINT_COUNT];
 			float a, b, m, n1, n2, n3;
 			float red, grn, blu;
-			float rotDirection;
+			int rotDirection;
 			float radiusMultiplier;
 		};
